Update prediction form styles
I changed the layout of this panel quite a bit from the mockups. The multiple steps in the original design of the form seemed unecessary given the amount of space we have in this panel now that the description is gone.

This is only the style updates, nothing is hooked up yet

Interested to hear what you all think…

# Conflicts:
#	app/components/MakePredictionForm.js