update decision data when buy/redeem events are fired
checks to see if this is a buy/redeem event that happened after the initial app load, and update decision data in app state. this ensures that the correct data will be there when the next buyMarketPositions transaction is constructed by the UI, minimizing the chance of an unexpected revert.

fixes https://github.com/levelkdev/futarchy-app/issues/121